Move Walker to tree.go
